+++ lib/class/api.class.php
@@ -2788,7 +2788,7 @@ class Api
         }
 
         if ($user_id > 0) {
-            if ($password) {
+            if ($password && !AmpConfig::get('simple_user_mode')) {
                 $user->update_password('', $password);
             }
             if ($fullname) {