+++ templates/show_search_bar.inc.php
@@ -33,7 +33,7 @@ $t_search = T_('Search'); ?>
             <option value="album"><?php echo $t_albums?></option>
             <option value="artist"><?php echo $t_artists?></option>
             <option value="playlist_name"><?php echo $t_playlists?></option>
-            <option value="tag"><?php echo T_('Tags')?></option>
+            <option value="tag"><?php echo T_('Genres')?></option>
             <?php if (AmpConfig::get('label')) { ?>
                 <option value="label"><?php echo T_('Labels')?></option>
             <?php